% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_criteria.R
\name{counts.per.seg}
\alias{counts.per.seg}
\title{Average count per segment for each replicat}
\usage{
counts.per.seg(list.tau, list.rna.data)
}
\arguments{
\item{list.tau}{A list of changepoints, for each dataset, for a given penalty.}

\item{list.rna.data}{A list of dataset (RNA countings)}
}
\value{
a matrix of average counts per segmentation per replica. The last line is the actual average counts per segment.
}
\description{
Average count per segment for each replicat
}
\examples{
library(robseg)
log.d1        <- log.transform(dataset1)
log.d2        <- log.transform(dataset2)
my.list.data  <- list(log.d1,log.d2)
rob_seg1      <- Rob_seg.std( x = log.d1 , loss = "Outlier", lambda = 25*log(length(dataset1)), lthreshold=3)
rob_seg2      <- Rob_seg.std( x = log.d2 , loss = "Outlier", lambda = 25*log(length(dataset1)), lthreshold=3)
my.list.tau   <- list(rob_seg1$t.est,rob_seg2$t.est)
cps      <- counts.per.seg(list.tau = my.list.tau,list.rna.data = my.list.data)
}
