% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_criteria.R
\name{log.transform}
\alias{log.transform}
\title{log-transformation function for segmentation}
\usage{
\method{log}{transform}(rna.data)
}
\arguments{
\item{rna.data}{: a vector of counts from RNA-sequencing}
}
\value{
a vector of log-transformed data
}
\description{
log-transformation function for segmentation
}
\examples{
log.data <- log.transform(dataset1)
plot(dataset1, type="l")
plot(log.data, type="l")
}
