% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_criteria.R
\name{mean.seg}
\alias{mean.seg}
\title{Segment mean per data}
\usage{
\method{mean}{seg}(tau, rna.data)
}
\arguments{
\item{tau}{a vector of changepoints}

\item{rna.data}{a vector of counts}
}
\value{
a vector of means per data
}
\description{
Segment mean per data
}
\examples{
library(robseg)
log.trans <- log.transform(dataset1)
seg_rob  <- Rob_seg.std(x = log.trans,  loss = "Outlier", lambda = beta, lthreshold=3)
tau       <- seg_rob$t.est
av.line   <- mean.seg(rna.data = data, tau = tau)
plot(av.line)

}
