% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_criteria.R
\name{mse.func}
\alias{mse.func}
\title{Computing the MSE from a pair of datasets, depending on a given penalty}
\usage{
mse.func(rna.data1, rna.data2, beta)
}
\arguments{
\item{rna.data1}{a vector of data}

\item{rna.data2}{a vector of data, same length as rna.data1}

\item{beta}{a value of penalty}
}
\value{
the value of MSE from two datasets by a given penalty
}
\description{
Computing the MSE from a pair of datasets, depending on a given penalty
}
\examples{
log.dataset1 <- log.transform(dataset1)
log.dataset2 <- log.transform(dataset2)
mse.val <- mse.func(rna.data1=log.dataset1,rna.data2=log.dataset2,beta=25*log(length(dataset1)))


}
