% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_criteria.R
\name{mse.penalties}
\alias{mse.penalties}
\title{Segmentation selection on MSE criterion}
\usage{
mse.penalties(list.rna.data, penalty_range)
}
\arguments{
\item{list.rna.data}{a list of vector of dataset}

\item{penalty_range}{a vector of length 2, respectively defining the minimum penalty and the maximum penalty}
}
\value{
A list with a vector of penalties and a matrix of mse
}
\description{
Segmentation selection on MSE criterion
}
\examples{
l.d1    <- log.transform(dataset1)
l.d2    <- log.transform(dataset2)
l.d3    <- log.transform(dataset3)
l.data  <- list(l.d1,l.d2,l.d3)
mse.res <- mse.penalties(list.rna.data=l.data, penalty_range=c(10,100))


}
