% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_criteria.R
\name{nid.func}
\alias{nid.func}
\title{NID computed from two changepoint datasets}
\usage{
nid.func(rna.data1, rna.data2, beta)
}
\arguments{
\item{rna.data1}{a vector of data}

\item{rna.data2}{a vector of data}

\item{beta}{penalty value}
}
\value{
the NID from the 2 segmentations, a decimal value between 0 and 1.
}
\description{
NID computed from two changepoint datasets
}
\examples{
log.dataset1 <- log.transform(dataset1)
log.dataset2 <- log.transform(dataset2)
nid.val <- nid.func(rna.data1=log.dataset1,rna.data2=log.dataset2,beta=25*log(length(dataset1)))

}
