% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_criteria.R
\name{plot.MSE}
\alias{plot.MSE}
\title{Plotting MSE object}
\usage{
\method{plot}{MSE}(mses)
}
\arguments{
\item{mse}{: the MSE objet returned by the function mse.penalties}
}
\value{
the graph according to the parameters entered in the function
}
\description{
Plotting MSE object
}
\examples{
mse.beta <- mse.penalties(list(log.transform(dataset1),log.transform(dataset2),log.transform(dataset3)), penalty_range = c(15,75) )
plot(mse.beta)

}
